+++ aether/node/src/main/java/org/pragmatica/aether/api/ManagementServer.java
@@ -13,6 +13,9 @@ import org.pragmatica.aether.api.routes.MetricsRoutes;
 import org.pragmatica.aether.api.routes.NodeLifecycleRoutes;
 import org.pragmatica.aether.api.routes.ObservabilityRoutes;
 import org.pragmatica.aether.api.routes.RepositoryRoutes;
+import org.pragmatica.aether.api.routes.ABTestRoutes;
+import org.pragmatica.aether.api.routes.BlueGreenRoutes;
+import org.pragmatica.aether.api.routes.CanaryRoutes;
 import org.pragmatica.aether.api.routes.RollingUpdateRoutes;
 import org.pragmatica.aether.api.routes.RouteHandler;
 import org.pragmatica.aether.api.routes.ScheduledTaskRoutes;
@@ -207,6 +210,9 @@ class ManagementServerImpl implements ManagementServer {
         routeSources.add(SliceRoutes.sliceRoutes(nodeSupplier));
         routeSources.add(MetricsRoutes.metricsRoutes(nodeSupplier, observability));
         routeSources.add(RollingUpdateRoutes.rollingUpdateRoutes(nodeSupplier));
+        routeSources.add(CanaryRoutes.canaryRoutes(nodeSupplier));
+        routeSources.add(BlueGreenRoutes.blueGreenRoutes(nodeSupplier));
+        routeSources.add(ABTestRoutes.abTestRoutes(nodeSupplier));
         routeSources.add(NodeLifecycleRoutes.nodeLifecycleRoutes(nodeSupplier));
         routeSources.add(RepositoryRoutes.repositoryRoutes(nodeSupplier));
         routeSources.add(ScheduledTaskRoutes.scheduledTaskRoutes(scheduledTaskRegistry,